--- mamas-search/src/main/java/com/sondahum/mamas/bid/domain/BidSearchFilter.java
@@ -1,4 +1,0 @@
-package com.sondahum.mamas.bid.domain;
-
-public enum BidSearchFilter {
-}
