--- mamas-search/src/main/java/com/sondahum/mamas/contract/dao/ContractRepository.java
@@ -1,9 +1,0 @@
-package com.sondahum.mamas.contract.dao;
-
-import com.sondahum.mamas.contract.domain.Contract;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ContractRepository extends JpaRepository<Contract, Long> {
-}
