--- mamas-search/src/main/java/com/sondahum/mamas/contract/domain/ContractSearchFilter.java
@@ -1,4 +1,0 @@
-package com.sondahum.mamas.contract.domain;
-
-public enum ContractSearchFilter {
-}
