+++ mamas-search/src/main/java/com/sondahum/mamas/repository/BidRepository.java
@@ -1,0 +1,15 @@
+package com.sondahum.mamas.repository;
+
+import com.sondahum.mamas.common.model.Action;
+import com.sondahum.mamas.domain.Bid;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface BidRepository extends JpaRepository<Bid, Long> {
+
+    List<Bid> findByAction(Action action);
+
+}
