+++ mamas-search/src/main/java/com/sondahum/mamas/repository/ContractRepository.java
@@ -1,0 +1,9 @@
+package com.sondahum.mamas.repository;
+
+import com.sondahum.mamas.domain.Contract;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ContractRepository extends JpaRepository<Contract, Long> {
+}
