+++ mamas-search/src/main/java/com/sondahum/mamas/repository/EstateRepository.java
@@ -1,0 +1,13 @@
+package com.sondahum.mamas.repository;
+
+import com.sondahum.mamas.domain.Estate;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface EstateRepository extends JpaRepository<Estate, Long> {
+
+    Optional<Estate> findByName(String name);
+}
