--- mamas-search/src/main/java/com/sondahum/mamas/user/domain/UserSearchQuery.java
@@ -1,4 +1,0 @@
-package com.sondahum.mamas.user.domain;
-
-public class UserSearchQuery {
-}
