+++ mamas-search/src/test/java/com/sondahum/mamas/repository/EstateRepositoryTest.java
@@ -1,10 +1,8 @@
 package com.sondahum.mamas.repository;
 
 import com.sondahum.mamas.AbstractMamasSearchTest;
-import com.sondahum.mamas.estate.domain.Estate;
-import com.sondahum.mamas.estate.dao.EstateRepository;
-import com.sondahum.mamas.user.dao.UserRepository;
-import com.sondahum.mamas.user.domain.User;
+import com.sondahum.mamas.domain.Estate;
+import com.sondahum.mamas.domain.User;
 import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
