+++ mamas-search/src/test/java/com/sondahum/mamas/repository/UserRepositoryTest.java
@@ -1,12 +1,7 @@
 package com.sondahum.mamas.repository;
 
 import com.sondahum.mamas.AbstractMamasSearchTest;
-import com.sondahum.mamas.bid.domain.Bid;
-import com.sondahum.mamas.bid.dao.BidRepository;
-import com.sondahum.mamas.estate.domain.Estate;
-import com.sondahum.mamas.estate.dao.EstateRepository;
-import com.sondahum.mamas.user.domain.User;
-import com.sondahum.mamas.user.dao.UserRepository;
+import com.sondahum.mamas.domain.User;
 import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
