+++ mamas-search/src/test/java/com/sondahum/mamas/service/CustomerManagerTest.java
@@ -1,8 +1,8 @@
 package com.sondahum.mamas.service;
 
 
-import com.sondahum.mamas.bid.dao.BidRepository;
-import com.sondahum.mamas.user.dao.UserRepository;
+import com.sondahum.mamas.repository.BidRepository;
+import com.sondahum.mamas.repository.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
